+++ src/main/java/com/xxmicloxx/NoteBlockAPI/songplayer/EntitySongPlayer.java
@@ -54,7 +54,7 @@ public class EntitySongPlayer extends RangeSongPlayer {
     }
 
     @Override
-    public void playTick(Player player, int tick) {
+    public void playTick(Player player, long tick) {
         if (!(currentPlaying instanceof Song))
             throw new IllegalStateException("Unexpected call to playTick");
         if (entity.isDead()) {
@@ -71,7 +71,7 @@ public class EntitySongPlayer extends RangeSongPlayer {
         byte playerVolume = NoteBlockAPI.getPlayerVolume(player);
 
         for (Layer layer : ((Song) currentPlaying).getLayerHashMap().values()) {
-            Note note = layer.getNote(tick);
+            Note note = layer.getNote((int) tick);
             if (note == null) continue;
 
             float volume = ((layer.getVolume() * (int) this.volume * (int) playerVolume * note.getVelocity()) / 100_00_00_00F)
